Added some more documentation and collected loose scripts into larger files
